requirements.txt
fastapi
uvicorn
jinja2
python-multipart
numpy==1.26.4
faiss-cpu==1.8.0
transformers==4.41.0
sentence-transformers==2.7.0
pyserini==0.21.0
pandas
scikit-learn
scipy
